Tests immediately crash.

+++ include/jaguar/diff_drive.h
@@ -22,9 +22,9 @@ public:
 private:
     virtual void init(uint16_t status_ms);
     virtual void heartbeat(void);
-    virtual void update_encoders_left(uint32_t pos);
-    virtual void update_encoders_right(uint32_t pos);
-    virtual void update_temperature(uint16_t temp);
+    virtual void update_encoders_left(double pos);
+    virtual void update_encoders_right(double pos);
+    virtual void update_temperature(double temp);
     virtual void update_estop(uint8_t state);
 
     virtual void block(can::TokenPtr t1, can::TokenPtr t2);
